Reduces duplication in backend specs
These specs were all being applied only to specific backends. Several of them are (functionally) the same, and do not need to be maintained separately. Those tests which do not apply to all backends are now clearly indicated by calling `test_only [some backend(s)]`, forming a todo list for building feature parity.
