+++ spec/lib/message_bus_spec.rb
@@ -250,38 +250,38 @@ describe MessageBus do
 
   end
 
-  unless MESSAGE_BUS_CONFIG[:backend] == :memory
-    it "should support forking properly do" do
-      data = []
-      @bus.subscribe do |msg|
-        data << msg.data
-      end
+  it "should support forking properly do" do
+    test_never :memory
 
-      @bus.publish("/hello", "pre-fork")
-      wait_for(2000) { data.length > 0 }
-
-      fork_has_published = false
-      child = Process.fork
-
-      if child
-        # The child was forked and we received its PID
-
-        # Wait for fork to finish so we're asserting that we can still publish after it has
-        Process.wait(child)
-
-        @bus.publish("/hello", "continuation")
-      else
-        begin
-          @bus.after_fork
-          @bus.publish("/hello", "from-fork")
-        ensure
-          exit!(0)
-        end
-      end
-
-      wait_for(2000) { data.length == 3 }
-
-      data.must_equal(["pre-fork", "from-fork", "continuation"])
+    data = []
+    @bus.subscribe do |msg|
+      data << msg.data
     end
+
+    @bus.publish("/hello", "pre-fork")
+    wait_for(2000) { data.length > 0 }
+
+    fork_has_published = false
+    child = Process.fork
+
+    if child
+      # The child was forked and we received its PID
+
+      # Wait for fork to finish so we're asserting that we can still publish after it has
+      Process.wait(child)
+
+      @bus.publish("/hello", "continuation")
+    else
+      begin
+        @bus.after_fork
+        @bus.publish("/hello", "from-fork")
+      ensure
+        exit!(0)
+      end
+    end
+
+    wait_for(2000) { data.length == 3 }
+
+    data.must_equal(["pre-fork", "from-fork", "continuation"])
   end
 end
