+++ spec/spec_helper.rb
@@ -30,3 +30,13 @@ def wait_for(timeout_milliseconds = 2000)
   end.join
 
 end
+
+def test_only(*backends)
+  backend = MESSAGE_BUS_CONFIG[:backend]
+  skip "Test doesn't apply to #{backend}" unless backends.include?(backend)
+end
+
+def test_never(*backends)
+  backend = MESSAGE_BUS_CONFIG[:backend]
+  skip "Test doesn't apply to #{backend}" if backends.include?(backend)
+end
